Review fer justin

+++ src/main/java/controllers/fads/FadsDeleteServlet.java
@@ -1,7 +1,8 @@
-package controllers;
+package controllers.fads;
 
 import dao.DaoFactory;
-import dao.Fads;
+import dao.fads.Fads;
+import services.Auth;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -15,14 +16,16 @@ import java.sql.SQLException;
 public class FadsDeleteServlet extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
+        Auth auth = new Auth(request);
+        if (!auth.shouldRedirect()) {
+            response.sendRedirect("/login");
+            return;
+        }
+
         long id = Long.parseLong(request.getParameter("id"));
 
-        try {
-            Fads fadsDao = DaoFactory.getFadsDao();
-            fadsDao.delete(id);
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
+        Fads fadsDao = DaoFactory.getFadsDao();
+        fadsDao.delete(id);
 
         response.sendRedirect("/fads");
